# openapi.yaml
openapi: 3.1.0
info:
  title: HWPX Skill API
  version: 0.3.0
  description: Inspect, search, validate, edit, and rebuild HWPX documents through GPT Actions.
servers:
  - url: https://hwpxskill-api.onrender.com
paths:
...
  /action/create-rich:
    post:
      operationId: createRichHwpx
      summary: Create a professional HWPX with tables and PNG charts.
      description: Builds and validates the HWPX. Present download_url directly to the user as a clickable link; do not fetch it through an Action.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRichHwpxRequest"
      responses:
        "200":
          description: Rich HWPX generation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RichActionResponse"
components:
  schemas:
...
    CreateRichHwpxRequest:
      type: object
      description: Markdown for a new or rebuilt rich HWPX. Use ```chart fences for server-rendered PNG charts.
      properties:
        filename: {type: string, minLength: 6, default: report.hwpx}
        markdown:
          type: string
          minLength: 1
          description: Full Markdown document. May contain Markdown tables and ```chart fences.
        preset:
          type: string
          enum: [보고서, 기안문, 계획서, 통지, 회의록, 개조식, 보도자료]
          default: 보고서
      required: [markdown]
      additionalProperties: false
...
    RichActionResponse:
      type: object
      properties:
        ok: {type: boolean}
        validated: {type: boolean}
        layout_warning_count: {type: integer}
        filename: {type: [string, "null"]}
        download_url:
          type: [string, "null"]
          format: uri
          description: Present this URL directly to the user as a clickable link.
        error_stage: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
      required: [ok, validated, layout_warning_count]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer